Move SqlHelper out of GraphStructures
We need it in several places now.
